Todo hasta el 25 de junio del 2025

+++ 06-06val_Array/Ejercicio1.cpp
@@ -17,6 +17,6 @@ int main(void){
 void print(std::string msg, const std::valarray<double>& v)
 {
     std::cout << msg << " ";
-    for (int n: v) { std::cout << std::setw(3) << n; }
+    for (double n: v) { std::cout << std::setw(3) << n; }
     std::cout << '\n';
 }
